requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Werkzeug==2.2.3
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
pymongo==4.6.0
Pillow==10.2.0